login test with page object, registration test - updates according to comments
