+++ kutsProj/src/test/java/loginTest/LoginTestWithPageObject.java
@@ -21,16 +21,4 @@ public class LoginTestWithPageObject extends BaseTest {
         checkExpectedResult("Button SignIn is not visible", loginPage.isButtonSignInPresent(), true);
         checkExpectedResult("Alert Invalid Username / password is not displayed", loginPage.isAlertInvalidUsernamePasswordDisplayed(), true);
     }
-
-    @Test
-    public void registrationAlerts(){
-        loginPage.openLoginPage();
-        loginPage.enterLoginInRegistration("tr");
-        loginPage.enterEmailRegistration("test.com");
-        loginPage.enterPassWordInRegistration("123");
-        loginPage.clickOnButtonSignUp();
-        checkExpectedResult("Alert Invalid Username Registration is not displayed", loginPage.isAlertInvalidUsernameRegistrationDisplayed(), true);
-        checkExpectedResult("Alert Invalid Email Registration is not displayed", loginPage.isAlertInvalidEmailRegistrationDisplayed(), true);
-        checkExpectedResult("Alert Invalid Password Registration is not displayed", loginPage.isAlertInvalidPasswordRegistrationDisplayed(), true);
-    }
 }
